+++ backend/src/controllers/user.controller.ts
@@ -1,5 +1,4 @@
 import { NextFunction, Request, Response } from "express";
-import { generateOTP, verifyOTP } from "../utils/twilitoHandller";
 import path from "path";
 import fs from "fs";
 import XLSX from "xlsx";
@@ -92,32 +91,6 @@ export async function getFilter(req: Request, res: Response): Promise<void> {
     const uniqueValues = [...new Set(data.map((row) => row[column]))].sort();
     res.status(200).send({ uniqueValues });
   } catch (e: any) {
-    res.status(500).send({ message: e.message });
-  }
-}
-
-export async function generateOtp(req: Request, res: Response): Promise<void> {
-  try {
-    console.log("Generating OTP");
-    const { phoneNumber } = req.body;
-    await generateOTP(phoneNumber);
-    res.status(200).send({ message: "OTP sent" });
-  } catch (e: any) {
-    console.log(e)
-    res.status(500).send({ message: e.message });
-  }
-}
-
-export async function verifyOtp(req: Request, res: Response): Promise<void> {
-  try {
-    console.log("Verifying OTP")
-    const { phoneNumber, code } = req.body;
-    console.log(phoneNumber, code)
-    const verified = await verifyOTP(phoneNumber, code);
-    if (!verified) throw Error("Invalid OTP");
-    res.status(200).send({ message: "OTP verified" });
-  } catch (e: any) {
-    console.log(e)
     res.status(500).send({ message: e.message });
   }
 }
