+++ backend/src/routes/userRoutes.ts
@@ -1,12 +1,13 @@
 import { Router } from "express";
-import { generateOtp, verifyOtp, getFilter, postRead } from "../controllers/user.controller";
+import { getFilter, postRead } from "../controllers/user.controller";
+import { generateOtp, verifyOtp } from "../controllers/admin.controller";
 import isAuth from "../middlewares/isClient";
 
 const userRouter: Router = Router();
 
 userRouter.post("/read", isAuth, postRead);
 userRouter.get("/filter", isAuth, getFilter);
-userRouter.post("/generate-otp", generateOtp)
-userRouter.post("/verify-otp", verifyOtp)
+userRouter.get("/generate-otp", generateOtp)
+userRouter.get("/verify-otp", verifyOtp)
 
 export default userRouter;
